requirements.txt
asgiref==3.8.1
blinker==1.8.2
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
cryptography==42.0.7
dj-database-url==2.2.0
Django==4.2.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
whitenoise==6.6.0
